# src/express-server/api/openapi.yaml
openapi: 3.0.0
info:
  title: Food Delivery API
  version: 1.0.0
  description: API REST para sistema de delivery - Trabalho DW1 UMAIA 2025/26
  contact:
    name: inf25dw1g13
    email: a046756@umaia.pt
components:
  schemas:
    Restaurante:
      title: Restaurante
      type: object
      required:
        - nome
        - morada
        - telefone
        - especialidade
      additionalProperties: false
      properties:
        id:
          type: integer
          readOnly: true
          description: ID do restaurante (gerado automaticamente)
        nome:
          type: string
          minLength: 3
          maxLength: 200
          description: Nome do restaurante
          example: "Pizzaria Napolitana"
        morada:
          type: string
          description: Morada completa do restaurante
          example: "Rua de Santa Catarina, 123"
        codigo_postal:
          type: string
          description: "(Opcional) Código postal"
          example: "4000-450"
        cidade:
          type: string
          description: "(Opcional) Cidade"
          example: "Porto"
        telefone:
          type: string
          description: "Número de telefone"
          example: "912345678"
        email:
          type: string
          format: email
          description: "(Opcional) Email de contacto"
          example: "info@pizzarianapolitana.pt"
        especialidade:
          type: string
          description: "Tipo de cozinha (ex: Italiana, Chinesa, Hambúrgueres)"
          example: "Italiana"
        descricao:
          type: string
          description: "(Opcional) Descrição do restaurante"
          example: "Pizza autêntica napolitana com forno a lenha"
        taxa_entrega:
          type: number
          format: float
          description: "(Opcional) Taxa de entrega em euros (padrão: 0.00)"
          example: 2.50
        pedido_minimo:
          type: number
          format: float
          description: "(Opcional) Valor mínimo do pedido em euros (padrão: 0.00)"
          example: 12.00
        ativo:
          type: boolean
          default: true
          description: "(Opcional) Se o restaurante está ativo (padrão: true)"
          example: true
    Prato:
      title: Prato
      type: object
      required:
        - restaurante_id
        - nome
        - preco
      additionalProperties: false
      properties:
        id:
          type: integer
          readOnly: true
          description: ID do prato (gerado automaticamente)
        restaurante_id:
          type: integer
          description: ID do restaurante a que pertence o prato
          example: 1
        nome:
          type: string
          description: Nome do prato
          example: "Pizza Margherita"
        descricao:
          type: string
          description: "(Opcional) Descrição do prato"
          example: "Pizza clássica com tomate, mozzarella e manjericão"
        preco:
          type: number
          format: float
          description: "Preço do prato em euros"
          example: 8.50
        vegetariano:
          type: boolean
          default: false
          description: "(Opcional) Se o prato é vegetariano (padrão: false)"
          example: true
        disponivel:
          type: boolean
          default: true
          description: "(Opcional) Se o prato está disponível (padrão: true)"
          example: true
...
    Error:
      title: Error
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
      # Schema é reutilizado em várias respostas de erro .
      # As respostas usam $ref para apontar para este modelo, todas as mensagens ficam com o mesmo formato
  examples:
    RestauranteExample01:
      value:
        nome: "Pizzaria Napolitana"
        morada: "Rua de Santa Catarina, 123"
        codigo_postal: "4000-450"
        cidade: "Porto"
        telefone: "912345678"
        email: "info@pizzarianapolitana.pt"
        especialidade: "Italiana"
        descricao: "Pizza autêntica napolitana com forno a lenha"
        taxa_entrega: 2.50
        pedido_minimo: 12.00
        ativo: true
    RestauranteExample02:
      value:
        nome: "Burger House"
        morada: "Avenida da Boavista, 500"
        telefone: "987654321"
        especialidade: "Hambúrgueres"
        taxa_entrega: 3.00
        pedido_minimo: 10.00
    PratoExample01:
      value:
        restaurante_id: 1
        nome: "Pizza Margherita"
        descricao: "Pizza clássica com tomate, mozzarella e manjericão"
        preco: 8.50
        vegetariano: true
        disponivel: true
    PratoExample02:
      value:
        restaurante_id: 2
        nome: "Burger Clássico"
        descricao: "Hambúrguer de carne, queijo, alface, tomate e molho especial"
        preco: 6.90
        vegetariano: false
        disponivel: true
...
paths:
  /api/restaurantes/{id}:
    get:
      tags: [Restaurantes]
      x-eov-operation-handler: controllers/RestaurantesController
      operationId: getRestaurante
      summary: Obter restaurante por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          examples:
            one:
              value: 1
              summary: Retrieve Restaurante id 1
            two:
              value: 2
              summary: Retrieve Restaurante id 2
      responses:
        '200':
          description: Restaurante encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurante'
        '404':
          description: Não encontrado
    put:
      tags: [Restaurantes]
      x-eov-operation-handler: controllers/RestaurantesController
      operationId: updateRestaurante
      summary: Atualizar restaurante
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          examples:
            one:
              value: 1
              summary: Update Restaurante id 1
            two:
              value: 2
              summary: Update Restaurante id 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Restaurante'
            examples:
              RestauranteExample01:
                $ref: '#/components/examples/RestauranteExample01'
              RestauranteExample02:
                $ref: '#/components/examples/RestauranteExample02'
      responses:
        '200':
          description: Atualizado
    delete:
      tags: [Restaurantes]
      x-eov-operation-handler: controllers/RestaurantesController
      operationId: deleteRestaurante
      summary: Apagar restaurante
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          examples:
            one:
              value: 1
              summary: Delete Restaurante id 1
            two:
              value: 2
              summary: Delete Restaurante id 2
      responses:
        '204':
          description: Removido
  /api/restaurantes/{id}/pratos:
    get:
      tags: [Restaurantes]
      x-eov-operation-handler: controllers/RestaurantesController
      operationId: getPratosRestaurante
      summary: Listar pratos de um restaurante
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          examples:
            one:
              value: 1
              summary: Listar pratos do restaurante id 1
            two:
              value: 2
              summary: Listar pratos do restaurante id 2
      responses:
        '200':
          description: Lista de pratos do restaurante
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prato'
        '404':
          description: Restaurante não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/restaurantes:
    get:
      tags: [Restaurantes]
      x-eov-operation-handler: controllers/RestaurantesController
      operationId: getRestaurantes
      summary: Listar restaurantes
      parameters:
        - name: especialidade
          in: query
          schema:
            type: string
          example: "Italiana"
          description: Filtrar por especialidade (opcional)
        - name: limite
          in: query
          schema:
            type: integer
          description: Limitar número de resultados (opcional)
      responses:
        '200':
          description: Lista de restaurantes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Restaurante'
    post:
      tags: [Restaurantes]
      x-eov-operation-handler: controllers/RestaurantesController
      operationId: createRestaurante
      summary: Criar restaurante
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Restaurante'
            examples:
              RestauranteExample01:
                $ref: '#/components/examples/RestauranteExample01'
              RestauranteExample02:
                $ref: '#/components/examples/RestauranteExample02'
      responses:
        '201':
          description: Restaurante criado
  /api/pratos:
    get:
      tags: [Pratos]
      x-eov-operation-handler: controllers/PratosController
      operationId: getPratos
      summary: Listar pratos
      parameters:
        - name: restaurante_id
          in: query
          schema:
            type: integer
        - name: categoria_id
          in: query
          schema:
            type: integer
        - name: vegetariano
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de pratos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prato'
    post:
      tags: [Pratos]
      x-eov-operation-handler: controllers/PratosController
      operationId: createPrato
      summary: Criar prato
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Prato'
            examples:
              PratoExample01:
                $ref: '#/components/examples/PratoExample01'
              PratoExample02:
                $ref: '#/components/examples/PratoExample02'
      responses:
        '201':
          description: Prato criado
...